% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R, R/DrugFormulation.R
\name{batch_variability}
\alias{batch_variability}
\title{Assess batch-to-batch variability}
\usage{
batch_variability(formulation_data, parameter)

batch_variability(formulation_data, parameter)
}
\arguments{
\item{formulation_data}{A data frame containing formulation data.}

\item{parameter}{The parameter for which batch-to-batch variability is calculated.}
}
\value{
The batch-to-batch variability of the specified parameter.

The batch-to-batch variability of the specified parameter.
}
\description{
This function calculates the batch-to-batch variability of a specified parameter.

This function calculates the batch-to-batch variability of a specified parameter.
}
