% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R, R/DrugFormulation.R
\name{compare_means}
\alias{compare_means}
\title{Compare means across groups}
\usage{
compare_means(formulation_data, group_var, response_var)

compare_means(formulation_data, group_var, response_var)
}
\arguments{
\item{formulation_data}{A data frame containing formulation data.}

\item{group_var}{The variable defining the groups for comparison.}

\item{response_var}{The response variable to compare across groups.}
}
\value{
Results of the t-test comparing means across groups.

Results of the t-test comparing means across groups.
}
\description{
This function compares the means of a response variable across groups specified by group_var.

This function compares the means of a response variable across groups specified by group_var.
}
