% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R, R/DrugFormulation.R
\name{confidence_intervals}
\alias{confidence_intervals}
\title{Confidence intervals of drug release}
\usage{
confidence_intervals(formulation_data)

confidence_intervals(formulation_data)
}
\arguments{
\item{formulation_data}{A data frame containing formulation data.}
}
\value{
Confidence intervals for drug release.

Confidence intervals for drug release.
}
\description{
This function computes confidence intervals for drug release based on the provided formulation data.

This function computes confidence intervals for drug release based on the provided formulation data.
}
