% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R, R/DrugFormulation.R
\name{control_chart}
\alias{control_chart}
\title{Control chart for quality control}
\usage{
control_chart(formulation_data, parameter)

control_chart(formulation_data, parameter)
}
\arguments{
\item{formulation_data}{A data frame containing formulation data.}

\item{parameter}{The quality control parameter to monitor.}
}
\value{
A control chart for the specified quality control parameter.

A control chart for the specified quality control parameter.
}
\description{
This function generates a control chart for monitoring the quality control parameter over time.

This function generates a control chart for monitoring the quality control parameter over time.
}
