% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugFormulation.R
\name{histogram}
\alias{histogram}
\title{Generate histogram}
\usage{
histogram(formulation_data, x, bins = 20)
}
\arguments{
\item{formulation_data}{A data frame containing the formulation data.}

\item{x}{The name of the variable.}

\item{bins}{The number of bins for the histogram.}
}
\value{
A histogram.
}
\description{
This function generates a histogram to visualize the distribution of a variable.
}
\examples{
formulation_data <- data.frame(
  Drug_Release = rnorm(100, mean = 50, sd = 10)
)
histogram(formulation_data, "Drug_Release")
}
