% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugFormulation.R
\name{scatterplot}
\alias{scatterplot}
\title{Generate scatterplot}
\usage{
scatterplot(formulation_data, x, y)
}
\arguments{
\item{formulation_data}{A data frame containing the formulation data.}

\item{x}{The name of the x-variable.}

\item{y}{The name of the y-variable.}
}
\value{
A scatterplot.
}
\description{
This function generates a scatterplot to visualize the relationship between two variables.
}
\examples{
formulation_data <- data.frame(
  Excipient_Concentration = runif(100, min = 0, max = 1),
  Drug_Release = rnorm(100, mean = 50, sd = 10)
)
scatterplot(formulation_data, "Excipient_Concentration", "Drug_Release")
}
